backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
deprecation==2.1.0
flask==3.1.2
flask-cors==6.0.1
google-auth==2.33.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.9
pycountry==24.6.1
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
strenum==0.4.15
typing-extensions==4.15.0
typing-inspection==0.4.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
websockets==15.0.1
werkzeug==3.1.3
